refactor: centralize frontend API calls into clean, named functions
- Create services/api.ts with specific functions (createTournament, updateTournament, etc.)
- Move helper functions (baseFetch, parseApiResponse) to utils/api.ts
- Simplify config/api.ts to only contain configuration and types
- Replace raw fetchWithAuth calls with descriptive function names
- Improve code maintainability across tournament management and admin interfaces
