+++ frontend/src/config/api.ts
@@ -1,3 +1,4 @@
+// config/api.ts - API configuration and types
 const API_BASE: string =
   process.env.NODE_ENV === "production"
     ? window.location.origin // Use the full origin URL in production
@@ -17,70 +18,4 @@ export const failure = <T>(error: string): ApiResponse<T> => ({
   error,
 });
 
-// Core function that preserves the ApiResponse
-const fetchApiResponseWithAuth = async <T>(
-  endpoint: string,
-  options: RequestInit = {},
-): Promise<ApiResponse<T>> => {
-  const token = localStorage.getItem("token");
-
-  const headers = {
-    ...options.headers,
-    Authorization: `Bearer ${token}`,
-  };
-
-  const response = await fetch(`${API_BASE}${endpoint}`, {
-    ...options,
-    headers,
-  });
-
-  console.log("response", response);
-
-  // For non-2xx responses, still try to parse the ApiResponse
-  if (!response.ok) {
-    try {
-      const errorResponse = await response.json();
-      // If it's already an ApiResponse failure, return it
-      if (
-        errorResponse &&
-        "success" in errorResponse &&
-        !errorResponse.success
-      ) {
-        return errorResponse as ApiResponse<T>;
-      }
-      // Otherwise create a failure response
-      return failure(errorResponse.message || "Request failed");
-    } catch {
-      return failure("Request failed");
-    }
-  }
-
-  if (response.status === 204) {
-    return success(null as T);
-  }
-
-  const data = await response.json();
-
-  // If already an ApiResponse, return as-is
-  if (data && "success" in data) {
-    return data as ApiResponse<T>;
-  }
-
-  // Legacy endpoint - wrap in success
-  return success(data as T);
-};
-
-// Convenience function that throws on failure (existing behavior)
-const fetchWithAuth = async <T>(
-  endpoint: string,
-  options: RequestInit = {},
-): Promise<T> => {
-  const response = await fetchApiResponseWithAuth<T>(endpoint, options);
-  if (response.success) {
-    return response.data;
-  } else {
-    throw new Error(response.error);
-  }
-};
-
-export { API_BASE, fetchWithAuth, fetchApiResponseWithAuth };
+export { API_BASE };
