+++ frontend/src/utils/matchApi.ts
@@ -1,7 +1,7 @@
 import { MatchWithPlayers } from "@shared/types/admin";
 import { CurrentMatch } from "@shared/types/currentMatch";
 
-import { fetchUserOverlayEndpoint, postAuthenticatedApiEndpoint } from "./api";
+import { fetchUserOverlayEndpoint, postAuthenticatedApiEndpoint } from "../services/api";
 
 export const fetchCurrentMatch = async (
   userId: number,
